chore(types): move more utility types into util module
