+++ packages/types/lib/util.ts
@@ -29,3 +29,46 @@ export type PluginType = 'plugin';
  * The strings referring to all extension types.
  */
 export type ExtensionType = DriverType | PluginType;
+
+/**
+ * Converts a kebab-cased string into a camel-cased string.
+ */
+export type KebabToCamel<S extends string> = S extends `${infer P1}-${infer P2}${infer P3}`
+  ? `${Lowercase<P1>}${Uppercase<P2>}${KebabToCamel<P3>}`
+  : Lowercase<S>;
+
+/**
+ * Converts an object with kebab-cased keys into camel-cased keys.
+ */
+export type ObjectToCamel<T> = {
+  [K in keyof T as KebabToCamel<string & K>]: T[K] extends Record<string, any>
+    ? KeysToCamelCase<T[K]>
+    : T[K];
+};
+
+/**
+ * Converts an object or array to have camel-cased keys.
+ */
+export type KeysToCamelCase<T> = {
+  [K in keyof T as KebabToCamel<string & K>]: T[K] extends Array<any>
+    ? KeysToCamelCase<T[K][number]>[]
+    : ObjectToCamel<T[K]>;
+};
+
+/**
+ * Object `B` has all the keys as object `A` (even if those keys in `A` are otherwise optional).
+ */
+export type Associated<A extends object, B extends {[key in keyof Required<A>]: unknown}> = {
+  [Prop in keyof Required<A>]: B[Prop];
+};
+
+/**
+ * Given `string` `T`, this is a case-insensitive version of `T`.
+ */
+export type AnyCase<T extends string> = string extends T
+  ? string
+  : T extends `${infer F1}${infer F2}${infer R}`
+  ? `${Uppercase<F1> | Lowercase<F1>}${Uppercase<F2> | Lowercase<F2>}${AnyCase<R>}`
+  : T extends `${infer F}${infer R}`
+  ? `${Uppercase<F> | Lowercase<F>}${AnyCase<R>}`
+  : '';
